requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
marshmallow==3.20.1
celery==5.3.1
redis==5.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1